i/apparmor: move mocking of home and overlay conditions to osutil
This is where they actually belong, since having the mocking at the
osutil level guarantees that the mocking is effective for all packages
calling into osutil.
