+++ interfaces/apparmor/export_test.go
@@ -47,15 +47,6 @@ func MockUnloadProfiles(f func(fnames []string, cacheDir string) error) (restore
 	return r
 }
 
-// MockIsRootWritableOverlay mocks the real implementation of osutil.IsRootWritableOverlay
-func MockIsRootWritableOverlay(new func() (string, error)) (restore func()) {
-	old := isRootWritableOverlay
-	isRootWritableOverlay = new
-	return func() {
-		isRootWritableOverlay = old
-	}
-}
-
 // MockProcSelfExe mocks the location of /proc/self/exe read by setupSnapConfineGeneratedPolicy.
 func MockProcSelfExe(symlink string) (restore func()) {
 	old := procSelfExe
